Devel back up Merge2 (#2)
* Project tested, only exception handling, listing and application entities remaining. Formal testing also remaining.

* Added authentication for login, cors and changed mobile no to string

* Added exception handling, removed adding duplicates with the same primary key(incredibly scuffed) and simple responses.

* Added notification entity, and full functionality for project attributes status and privacy status.

* Fully implemented notifications, fixed a few bugs, refactored some things. Now only formal testing, constraints, application and listing remaining.

* Username uniqueness endpoint added

* Listing entity fully added, fixed a few bugs for deletion and restructured deletion. Shifted all integer primary keys to long for scalability.

* Application fully added but testing remains. Formal testing and adding constraints also remains.

+++ src/main/java/com/DevConnect/BE/Repo/ProjectRepo.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ProjectRepo extends JpaRepository<Project, Integer>
+public interface ProjectRepo extends JpaRepository<Project, Long>
 {
     List<Project> findByCategoryIn(List <String> Category);
 
@@ -18,5 +18,5 @@ public interface ProjectRepo extends JpaRepository<Project, Integer>
     List<Project> findByRoleContaining(String Role);
 
     @Query("select u.username from Project p join User u where p.id = ?1")
-    List<String> findAllCollaborator(Integer id);
+    List<String> findAllCollaborator(Long id);
 }
